Now use "priority" rather than "is_fallback"
This enables more fine-grained control of which
Reader/Writer module will be used for any given filetype.
Mainly important in regard to the Xattr module,
because that can be used with ANY file, so we want
that to be checked last.

Also changed things so that only one Reader will
be used, not all-possible-readers, because that
can give confusing and inconsistent results.
I mean, if only one Writer can be used, then only one
Reader should be used, one that matches the Writer.