Fixed bug with losing data!
If data had been set using the Xattr writer (in an earlier
version before Exif had been altered to accept any field)
then the bug was that it would fail to read that Xattr data
because it was ONLY using the Exif reader. Now it uses all
possible readers (like it used to do!) in priority order.
So Exif will take precedence, but Xattr will not be ignored.
Whew!