Fixing bug where soft links are destroyed
Basically ExifTool does NOT honour soft links!
It replaces a soft link with the contents of the file, thus
(a) destroying the soft link, and
(b) not actually updating the file you wanted to update.
While the "exiftool" script has the -overwrite_in_place option,
the Perl module Image::ExifTool does not.
But I figured out a workaround.